Timestamp fuzzer in metav1 should not use negative values
While debugging a similar issue in Cluster API, found that the Creation
and Deletion timestamp were mostly `nil`. After a bit of digging it
seems that using the Fuzz on an int64 can cause the Rfc3339Copy() to
error out (silently, the error is actually dropped) and return a nil
value. When using int64 large negative numbers can be generated
randomly, which result in an invalid date.

Signed-off-by: Vince Prignano <vincepri@vmware.com>

Kubernetes-commit: 96f7d91cc7664e00949fdb556b9a723bcf029e05